rename package to com.sample
