+++ src/main/resources/samples.conf
@@ -32,7 +32,7 @@ dataObjects {
   custom-my-df {
     type = CustomDfDataObject
     creator {
-      class-name = com.test.CreateMyDf
+      class-name = com.sample.CreateMyDf
     }
   }
 
@@ -116,7 +116,7 @@ dataObjects {
   custom-rating-df {
     type = CustomDfDataObject
     creator {
-      class-name = com.test.CreateRatingDf
+      class-name = com.sample.CreateRatingDf
     }
   }
 
@@ -206,7 +206,7 @@ actions {
     type = CopyAction
     inputId = ab-csv-org
     outputId = ab-reduced-hsqldb
-    transformer.class-name = com.test.ReduceNycCSVTransformer
+    transformer.class-name = com.sample.ReduceNycCSVTransformer
     metadata {
       feed = ab-jdbc
     }
@@ -241,7 +241,7 @@ actions {
     type = CustomSparkAction
     inputIds = [custom-rating-csv1, custom-rating-csv2]
     outputIds = [custom-rating-csv-agg]
-    transformer.class-name = com.test.RatingTransformer
+    transformer.class-name = com.sample.RatingTransformer
     metadata {
       feed = custom-rating-csv
     }
