+++ src/main/scala/com/sample/CreateMyDf.scala
@@ -1,4 +1,4 @@
-package com.test
+package com.sample
 
 import io.smartdatalake.workflow.action.customlogic.CustomDfCreator
 import org.apache.spark.sql.{DataFrame, SparkSession}
