+++ src/main/scala/com/sample/ReduceNycCSVTransformer.scala
@@ -1,4 +1,4 @@
-package com.test
+package com.sample
 
 import io.smartdatalake.workflow.action.customlogic.CustomDfTransformer
 import org.apache.spark.sql.{DataFrame, SparkSession}
